refactor

+++ service/src/main/java/com/github/al/realworld/application/command/DeleteArticleHandler.java
@@ -1,14 +1,17 @@
-package com.github.al.realworld.application.command.article;
+package com.github.al.realworld.application.command;
 
 import com.github.al.realworld.api.command.DeleteArticle;
 import com.github.al.realworld.api.command.DeleteArticleResult;
 import com.github.al.realworld.bus.CommandHandler;
-import com.github.al.realworld.domain.Article;
+import com.github.al.realworld.domain.model.Article;
 import com.github.al.realworld.domain.repository.ArticleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
+
+import static com.github.al.realworld.application.exception.NoAuthorizationException.forbidden;
 import static com.github.al.realworld.application.exception.ResourceNotFoundException.notFound;
 
 @RequiredArgsConstructor
@@ -22,6 +25,10 @@ public class DeleteArticleHandler implements CommandHandler<DeleteArticleResult,
     public DeleteArticleResult handle(DeleteArticle command) {
         Article article = articleRepository.findBySlug(command.getSlug())
                 .orElseThrow(() -> notFound("article [slug=%s] does not exist", command.getSlug()));
+
+        if (!Objects.equals(article.getAuthor().getUsername(), command.getCurrentUsername())) {
+            throw forbidden();
+        }
 
         articleRepository.delete(article);
